+++ src/main/scala/com/databricks/labs/overwatch/pipeline/Restore.scala
@@ -3,8 +3,11 @@ package com.databricks.labs.overwatch.pipeline
 import com.databricks.dbutils_v1.DBUtilsHolder.dbutils
 import com.databricks.labs.overwatch.env.{Database, Workspace}
 import com.databricks.labs.overwatch.utils.Helpers.{logger, removeTrailingSlashes}
-import com.databricks.labs.overwatch.utils.{CloneDetail, Config, Helpers, SparkSessionWrapper, WriteMode}
+import com.databricks.labs.overwatch.utils.{BadConfigException, CloneDetail, Config, Helpers, SparkSessionWrapper, WriteMode}
 import org.apache.log4j.{Level, Logger}
+import com.databricks.labs.overwatch.pipeline.Snapshot
+
+import java.io.FileNotFoundException
 
 class Restore (_sourceETLDB: String, _targetPrefix: String, _workspace: Workspace, _database: Database, _config: Config)
   extends Pipeline(_workspace, _database, _config) {
@@ -22,9 +25,8 @@ class Restore (_sourceETLDB: String, _targetPrefix: String, _workspace: Workspac
   val Config = _config
 
   private[overwatch] def restore(
-                                  globalPath: String,
+                                  allTarget: Array[PipelineTable],
                                   cloneLevel: String = "DEEP",
-                                  mode: String = "Overwrite",
                                 ): Unit = {
     val acceptableCloneLevels = Array("DEEP", "SHALLOW")
     require(acceptableCloneLevels.contains(cloneLevel.toUpperCase), s"SNAP CLONE ERROR: cloneLevel provided is " +
@@ -36,54 +38,37 @@ class Restore (_sourceETLDB: String, _targetPrefix: String, _workspace: Workspac
       case e: Exception => dbutils.fs.mkdirs(targetPrefix)
     }
 
-    if (mode.toLowerCase() == "overwrite") { // Restore in Overwrite Mode
+    try {
       if (!dbutils.fs.ls(targetPrefix).nonEmpty) {
-        val sourcesToSnap = workspace.getWorkspaceDatasets
-        val cloneSpecs = sourcesToSnap.map(dataset => {
-          val sourceName = dataset.name
-          val sourcePath = dataset.path
-          val targetPath = s"$targetPrefix/global_share/$sourceName"
-          CloneDetail(sourcePath, targetPath, None, cloneLevel)
-        }).toArray.toSeq
+        val cloneSpecs = new Snapshot(workSpace.getConfig.databaseName, targetPrefix, workspace, workspace.database, workspace.getConfig, "restore").buildCloneSpecs(cloneLevel, allTarget)
         val cloneReport = Helpers.parClone(cloneSpecs)
         val restoreReportPath = s"${targetPrefix}/restore_report/"
-        cloneReport.toDS.write.format("delta").mode("overwrite").save(restoreReportPath)
+        cloneReport.toDS.write.format("delta").mode("append").save(restoreReportPath)
       } else {
-        println("Target Path is not Empty...... Could not proceed with OverWrite Mode. Please select Append Mode of Restoration")
+        println("Target Path is not Empty...... Could not proceed with Restoration")
       }
-    } else { // Restore in Append Mode
-      try {
-        val allTargets = (bronze.getAllTargets ++ silver.getAllTargets ++ gold.getAllTargets).filter(_.exists(dataValidation = true))
-        val newTargets = allTargets.map(_.copy(_mode = WriteMode.append, withCreateDate = false, withOverwatchRunID = false))
-        newTargets.foreach(t => {
-          val tableName = t.name.toLowerCase()
-          println(s"Data is written from ${globalPath} to ${targetPrefix}/global_share/${tableName} with writeMode ${t.writeMode}")
-          val inputDF = spark.read.format("delta").load(s"${globalPath}/${tableName}")
-          inputDF.write.format("delta").mode(s"${t.writeMode}").option("mergeSchema", "true").save(s"${targetPrefix}/global_share/${tableName}")
-        })
 
-      } catch {
-        case e: Throwable =>
-          val failMsg = PipelineFunctions.appendStackStrace(e)
-          logger.log(Level.WARN, failMsg)
-      }
+    } catch {
+      case e: Throwable =>
+        val failMsg = PipelineFunctions.appendStackStrace(e)
+        logger.log(Level.WARN, failMsg)
     }
-
   }
-
 }
 
 object Restore extends SparkSessionWrapper {
+  private val logger: Logger = Logger.getLogger(this.getClass)
 
-    def apply(workspace: Workspace,
-              targetPrefix : String,
-              backupPath: String,
-              CloneLevel: String,
-              modeOfRestoration: String
+    def apply(
+               workSpace: Workspace,
+               allTarget: Array[PipelineTable],
+               targetPrefix : String,
+               backupPath: String,
+               CloneLevel: String
              ): Any = {
 
-      val restoration = new Restore(workspace.getConfig.databaseName, targetPrefix, workspace, workspace.database, workspace.getConfig)
-      restoration.restore(backupPath,"Deep",modeOfRestoration)
+      val restoration = new Restore(workSpace.getConfig.databaseName, targetPrefix, workSpace, workSpace.database, workSpace.getConfig)
+      restoration.restore(allTarget,CloneLevel)
 
     }
 
@@ -91,25 +76,54 @@ object Restore extends SparkSessionWrapper {
      * Create a backup of the Overwatch datasets
      *
      * @param arg(0)        Source ETL Path Prefix
-     * @param arg(1)        Source Organization ID
      * @param arg(2)        Target ETL Path Prefix
-     * @param arg(3)        Mode of Restore. If "Overwrite" then purge the target and restore otherwise append on the existing one.
      * @return
      */
 
     def main(args: Array[String]): Unit = {
 
       val sourcePrefix = args(0)
-      val orgID = args(1)
-      val targetPrefix = args(2)
-      val modeOfRestoration = args(3)
+      val orgID = Initializer.getOrgId
+      val targetPrefix = args(1)
+      val cloneLevel = "Deep"
+      //
 
-      val backupPath = s"${sourcePrefix}/data"  //As we are restoring from backup path  so path will be sourcePrefix/data/pipeline_report
+
+      val backupPath = s"${sourcePrefix}/data"
       val pipReportPath = s"${backupPath}/pipeline_report"
+      try{
+        dbutils.fs.ls(s"$pipReportPath/_delta_log").nonEmpty
+        logger.log(Level.INFO, s"Overwatch has being deployed with ${pipReportPath} location...proceed")
+      }catch {
+        case e: FileNotFoundException =>
+          val msg = s"Overwatch has not been deployed with ${pipReportPath} location...can not proceed"
+          logger.log(Level.ERROR, msg)
+          throw new BadConfigException(msg)
+      }
+
 
       val workSpace = Helpers.getRemoteWorkspaceByPath(pipReportPath,true,orgID)
+//
+//      val sourceETlDB = workSpace.getConfig.databaseName
+//      val snapshotRootPath = targetPrefix
+//      val pipeline = "Bronze,Silver,Gold"
+//      val cloneLevel = "Deep"
+//      val snapshotType = "full"
+//      val tablesToExclude = ""
+//      val processType = "Restore"
+//
+//      // Step 01 - Start Restore Process
+//      Snapshot.main(Array(sourceETlDB,snapshotRootPath,pipeline,snapshotType,tablesToExclude,cloneLevel,processType))
 
-      Restore(workSpace,targetPrefix,backupPath,"Deep",modeOfRestoration)
+      val bronze = Bronze(workSpace)
+      val silver = Silver(workSpace)
+      val gold = Gold(workSpace)
+      val pipelineReport = bronze.pipelineStateTarget
+      val allTarget = bronze.getAllTargets ++ silver.getAllTargets ++ gold.getAllTargets ++ Array(pipelineReport).filter(_.exists(dataValidation = true))
+
+      // Step 1 : Restore Process Started
+      Restore(workSpace,allTarget,targetPrefix,backupPath,cloneLevel)
+
 
       println("Restore Completed")
     }
